Combines data into test file and adds missing tests for pool members
